docs: augment write API sample to add more types
This PR adds some missing functionality to the appending data
example for node, namely to demonstrate writing complex messages
with nested message types.

It augments the sample data proto to be in line with other languages
and adds the testing to ensure that complex types are propagated
through to query results.